Flooding algorithm shouldn't have stack overflow errors anymore. We need to discuss "out of bounds" issues involving that algorithm and other functions. I also did'nt finish updating player score to a 16 bit uint because I was scared to mess up Joshua's stuff, but I'm pretty sure score is now implemented into gamelogic correctly.

+++ server/player.cpp
@@ -108,12 +108,12 @@ void Player::kill()
 	y = -2;
 }
 
-quint8 Player::getScore() const
+quint16 Player::getScore() const
 {
 	return score;
 }
 
-void Player::setScore(quint8 sc)
+void Player::setScore(quint16 sc)
 {
 	sc = sc > 200 ? 200 : sc;
 	if (sc == score)
